Update create_master_lists.py
as requested in :

- https://github.com/wmo-im/GRIB2/issues/219

>In GRIB2 Template definition files the octet number (from - to) for each entry in the GRIB section is given.
Most of the GRIB processing software packages need the length or number of octets for each entry, which has to be calculated from the specification of "OctetNo".
But with variables and repitition within some templates an automated calculation is sometimes not easy. (e.g. 37 + (ND-1)*4 + (NF-1)*4 -40 +(ND-1)*4 + (NF-1)*4)
Therefore, this is a proposal to add a column with the length of each entry in the GRIB2_Template files.