219 new (#307)
* Update create_master_lists.py

as requested in :

- https://github.com/wmo-im/GRIB2/issues/219

>In GRIB2 Template definition files the octet number (from - to) for each entry in the GRIB section is given.
Most of the GRIB processing software packages need the length or number of octets for each entry, which has to be calculated from the specification of "OctetNo".
But with variables and repitition within some templates an automated calculation is sometimes not easy. (e.g. 37 + (ND-1)*4 + (NF-1)*4 -40 +(ND-1)*4 + (NF-1)*4)
Therefore, this is a proposal to add a column with the length of each entry in the GRIB2_Template files.

* Update create_master_lists.py

remove any keys not in 'fieldnames' to avoid ValueError

* xml,txt files

* Update create_master_lists.py

make the count to be computed when the limit is not fixed but variable

* xml,txt files

* run the script update_octetcount.py to update the OctetCount number for all templates. the script will be added in the issue #219

* Replace Length per octetCount and remove it from the xml

* xml,txt files

* remove the octetCount in the txt and from the CI/CD

* update the create_master to avoid any issue while populating fields in the xml and txt

* update the create_master to avoid any issue while populating fields in the xml and txt

* Apply suggestions from code review

* xml,txt files

---------

Co-authored-by: antoineMerleEUM <antoine.merle@eumetsat.int>
Co-authored-by: Enrico Fucile <efucile@wmo.int>
Co-authored-by: antoinemerle <am92300@gmail.com>